+++ src/main/java/roomescape/dao/jdbc/ReservationJdbcDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.dao.jdbc;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,6 +9,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.dao.ReservationDao;
 import roomescape.domain.Reservation;
 
 @Repository
