+++ src/main/java/roomescape/dao/jdbc/ThemeJdbcDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.dao.jdbc;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -10,6 +10,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
 import roomescape.domain.vo.Name;
 import roomescape.dto.PopularThemeRequestDto;
