+++ src/main/java/roomescape/dao/jdbc/TimeJdbcDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.dao.jdbc;
 
 import java.time.LocalTime;
 import java.util.List;
@@ -9,6 +9,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.dao.TimeDao;
 import roomescape.domain.Time;
 
 @Repository
