+++ src/test/java/roomescape/dao/ThemeJdbcDaoTest.java
@@ -12,6 +12,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
+import roomescape.dao.jdbc.ReservationJdbcDao;
+import roomescape.dao.jdbc.ThemeJdbcDao;
+import roomescape.dao.jdbc.TimeJdbcDao;
 import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
 import roomescape.domain.Time;
