+++ src/test/java/roomescape/dao/TimeJdbcDaoTest.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
+import roomescape.dao.jdbc.TimeJdbcDao;
 import roomescape.domain.Time;
 
 @JdbcTest
